refactor: streamline OWA MCAP viewer architecture for maintainability
- Consolidate API routes from 4 files into single routes.py
- Implement proper service layer with dependency injection
- Create dedicated services/ folder with focused responsibilities:
  * CacheService: cache management only
  * FileService: file operations only
  * McapService: MCAP processing only
- Merge repositories into unified repository.py
- Consolidate models into single models.py file
- Unify utilities and exceptions into utils.py
- Simplify app.py and config.py structure

Benefits:
- Reduced Python files from 21 to 12 (43% reduction)
- Improved separation of concerns and testability
- Maintained 100% API compatibility
- Better code organization and maintainability
- Clean dependency injection pattern

All existing functionality preserved including file upload,
MCAP viewing, caching, and export capabilities.
