+++ projects/owa-mcap-viewer/owa_viewer/app.py
@@ -9,11 +9,9 @@ from rich.logging import RichHandler
 
 from .api import router as api_router
 from .config import settings
-from .services.cache_service import cache_service
-from .services.file_service import file_service
-from .utils.formatting import format_size_human_readable
+from .services import services
+from .utils import format_size_human_readable
 
-# Set up logging, use rich handler
 logging.basicConfig(
     level=settings.LOG_LEVEL,
     handlers=[RichHandler(rich_tracebacks=True, show_time=False, show_level=True)],
@@ -23,38 +21,21 @@ logger = logging.getLogger(__name__)
 
 @asynccontextmanager
 async def lifespan(app: FastAPI):
-    """Handle application lifespan events"""
+    """Handle application startup and shutdown"""
     logger.info("Starting OWA Viewer application")
-    # Run immediate clear
-    cache_service.clear()
-    # Start periodic cleanup (every hour by default)
-    await cache_service.start_periodic_cleanup()
-
+    services.clear_cache()
+    await services.start_periodic_cleanup()
     yield
-
     logger.info("Shutting down OWA Viewer application")
-    # Stop periodic cleanup
-    await cache_service.stop_periodic_cleanup()
+    await services.stop_periodic_cleanup()
 
 
 app = FastAPI(lifespan=lifespan)
-
-# Enable CORS
 app.add_middleware(
-    CORSMiddleware,
-    allow_origins=["*"],
-    allow_credentials=True,
-    allow_methods=["*"],
-    allow_headers=["*"],
+    CORSMiddleware, allow_origins=["*"], allow_credentials=True, allow_methods=["*"], allow_headers=["*"]
 )
-
-# Include the API router
 app.include_router(api_router)
-
-# Mount static files (CSS, JS)
 app.mount("/static", StaticFiles(directory="static"), name="static")
-
-# Setup Jinja2 templates
 templates = Jinja2Templates(directory="templates")
 
 
@@ -62,35 +43,20 @@ templates = Jinja2Templates(directory="templates")
 async def read_root(request: Request):
     """Render the index page"""
     featured_datasets = settings.FEATURED_DATASETS.copy()
-
     if settings.PUBLIC_HOSTING_MODE and "local" in featured_datasets:
         featured_datasets.remove("local")
-
-    return templates.TemplateResponse(
-        "index.html",
-        {
-            "request": request,
-            "featured_datasets": featured_datasets,
-        },
-    )
+    return templates.TemplateResponse("index.html", {"request": request, "featured_datasets": featured_datasets})
 
 
 @app.get("/viewer")
 async def read_viewer(repo_id: str, request: Request):
     """Render the viewer page for a specific dataset"""
-    # Get files from the repository
-    files = file_service.list_files(repo_id)
-
-    # Calculate total size and format it as human-readable
+    files = services.file_service.list_files(repo_id)
     total_size = sum(f.size for f in files) if files else 0
     size_str = format_size_human_readable(total_size)
-
     return templates.TemplateResponse(
         "viewer.html",
-        {
-            "request": request,
-            "dataset_info": {"repo_id": repo_id, "files": len(files), "size": size_str},
-        },
+        {"request": request, "dataset_info": {"repo_id": repo_id, "files": len(files), "size": size_str}},
     )
