+++ projects/owa-mcap-viewer/owa_viewer/config.py
@@ -4,29 +4,27 @@ from pathlib import Path
 from dotenv import load_dotenv
 from pydantic_settings import BaseSettings
 
-# Load environment variables
 load_dotenv()
 
 
 class Settings(BaseSettings):
-    """Centralized application settings"""
+    """Application settings"""
 
-    # File storage settings
+    # File storage
     EXPORT_PATH: str = os.environ.get("EXPORT_PATH", "./data")
     PUBLIC_HOSTING_MODE: bool = not bool(os.environ.get("EXPORT_PATH"))
 
-    # Cache settings
-    # MCAP_CACHE_SIZE_LIMIT: int = int(os.environ.get("MCAP_CACHE_SIZE_LIMIT", 1024 * 1024 * 1024))  # 1GB default, not implemented yet
-    FILE_CACHE_TTL: int = int(os.environ.get("FILE_CACHE_TTL", 600))  # 10 min default
-    DEFAULT_CACHE_TTL: int | None = int(os.environ.get("DEFAULT_CACHE_TTL", 3600))  # 1 hour default
+    # Cache
+    FILE_CACHE_TTL: int = int(os.environ.get("FILE_CACHE_TTL", 600))
+    DEFAULT_CACHE_TTL: int = int(os.environ.get("DEFAULT_CACHE_TTL", 3600))
     CACHE_DIR: str = os.environ.get("CACHE_DIR", "./cache")
 
-    # Server settings
+    # Server
     PORT: int = int(os.environ.get("PORT", 7860))
     HOST: str = os.environ.get("HOST", "0.0.0.0")
     LOG_LEVEL: str = os.environ.get("LOG_LEVEL", "INFO")
 
-    # File patterns
+    # Featured datasets
     FEATURED_DATASETS: list[str] = [
         "local",
         "open-world-agents/example_dataset",
@@ -41,6 +39,4 @@ class Settings(BaseSettings):
 
 
 settings = Settings()
-
-# Ensure cache directory exists
 Path(settings.CACHE_DIR).mkdir(parents=True, exist_ok=True)
